transformers/appairflow5/requirements.txt
apache-airflow
azure-storage-file-share
azure-storage-blob
pandas
pyyaml
